Move Consume back to application facade; rework mocks and tests
