+++ api/application/client.go
@@ -406,3 +406,25 @@ func (c *Client) DestroyRelation(endpoints ...string) error {
 	params := params.DestroyRelation{Endpoints: endpoints}
 	return c.facade.FacadeCall("DestroyRelation", params, nil)
 }
+
+// Consume adds a remote application to the model.
+func (c *Client) Consume(offer params.ApplicationOffer, alias string) (string, error) {
+	var consumeRes params.ConsumeApplicationResults
+	args := params.ConsumeApplicationArgs{
+		Args: []params.ConsumeApplicationArg{{
+			ApplicationAlias: alias,
+			ApplicationOffer: offer,
+		}},
+	}
+	err := c.facade.FacadeCall("Consume", args, &consumeRes)
+	if err != nil {
+		return "", errors.Trace(err)
+	}
+	if resultLen := len(consumeRes.Results); resultLen != 1 {
+		return "", errors.Errorf("expected 1 result, got %d", resultLen)
+	}
+	if err := consumeRes.Results[0].Error; err != nil {
+		return "", errors.Trace(err)
+	}
+	return consumeRes.Results[0].LocalName, nil
+}
