+++ apiserver/application/application_unit_test.go
@@ -14,19 +14,22 @@ import (
 	"github.com/juju/juju/apiserver/application"
 	"github.com/juju/juju/apiserver/params"
 	apiservertesting "github.com/juju/juju/apiserver/testing"
+	"github.com/juju/juju/environs"
+	"github.com/juju/juju/feature"
 	"github.com/juju/juju/juju"
+	"github.com/juju/juju/network"
 	"github.com/juju/juju/state"
 	coretesting "github.com/juju/juju/testing"
 )
 
 type ApplicationSuite struct {
 	testing.IsolationSuite
-	backend     mockBackend
-	application mockApplication
-	charm       mockCharm
-	endpoints   []state.Endpoint
-	relation    mockRelation
+	coretesting.JujuOSEnvSuite
+	backend   mockBackend
+	endpoints []state.Endpoint
+	relation  mockRelation
 
+	env          *mockEnviron
 	blockChecker mockBlockChecker
 	authorizer   apiservertesting.FakeAuthorizer
 	api          *application.API
@@ -34,40 +37,57 @@ type ApplicationSuite struct {
 
 var _ = gc.Suite(&ApplicationSuite{})
 
+func (s *ApplicationSuite) SetUpSuite(c *gc.C) {
+	s.SetInitialFeatureFlags(feature.CrossModelRelations)
+	s.IsolationSuite.SetUpSuite(c)
+}
+
 func (s *ApplicationSuite) SetUpTest(c *gc.C) {
 	s.IsolationSuite.SetUpTest(c)
+	s.JujuOSEnvSuite.SetUpTest(c)
 	s.authorizer = apiservertesting.FakeAuthorizer{
 		Tag: names.NewUserTag("admin"),
 	}
-	s.application = mockApplication{
-		units: []mockUnit{{
-			tag: names.NewUnitTag("foo/0"),
-		}, {
-			tag: names.NewUnitTag("foo/1"),
-		}},
-	}
-	s.charm = mockCharm{
-		config: &charm.Config{
-			Options: map[string]charm.Option{
-				"stringOption": {Type: "string"},
-				"intOption":    {Type: "int", Default: int(123)},
-			},
-		},
-	}
+	s.env = &mockEnviron{}
 	s.endpoints = []state.Endpoint{
-		{ApplicationName: "foo"},
+		{ApplicationName: "postgresql"},
 		{ApplicationName: "bar"},
 	}
 	s.relation = mockRelation{}
 	s.backend = mockBackend{
-		application: &s.application,
-		charm:       &s.charm,
-		endpoints:   &s.endpoints,
-		relation:    &s.relation,
+		applications: map[string]application.Application{
+			"postgresql": &mockApplication{
+				name: "postgresql",
+				charm: &mockCharm{
+					config: &charm.Config{
+						Options: map[string]charm.Option{
+							"stringOption": {Type: "string"},
+							"intOption":    {Type: "int", Default: int(123)},
+						},
+					},
+				},
+				units: []mockUnit{{
+					tag: names.NewUnitTag("postgresql/0"),
+				}, {
+					tag: names.NewUnitTag("postgresql/1"),
+				}},
+			},
+		},
+		remoteApplications: make(map[string]application.RemoteApplication),
+		charm: &mockCharm{
+			config: &charm.Config{
+				Options: map[string]charm.Option{
+					"stringOption": {Type: "string"},
+					"intOption":    {Type: "int", Default: int(123)},
+				},
+			},
+		},
+		endpoints: &s.endpoints,
+		relation:  &s.relation,
 		unitStorageAttachments: map[string][]state.StorageAttachment{
-			"foo/0": {
+			"postgresql/0": {
 				&mockStorageAttachment{
-					unit:    names.NewUnitTag("foo/0"),
+					unit:    names.NewUnitTag("postgresql/0"),
 					storage: names.NewStorageTag("pgdata/0"),
 				},
 			},
@@ -75,7 +95,7 @@ func (s *ApplicationSuite) SetUpTest(c *gc.C) {
 		storageInstances: map[string]*mockStorage{
 			"pgdata/0": {
 				tag:   names.NewStorageTag("pgdata/0"),
-				owner: names.NewUnitTag("foo/0"),
+				owner: names.NewUnitTag("postgresql/0"),
 			},
 		},
 	}
@@ -91,10 +111,14 @@ func (s *ApplicationSuite) SetUpTest(c *gc.C) {
 		func(application.Backend, juju.DeployApplicationParams) error {
 			return nil
 		},
-		nil,
 	)
 	c.Assert(err, jc.ErrorIsNil)
 	s.api = api
+}
+
+func (s *ApplicationSuite) TearDownTest(c *gc.C) {
+	s.JujuOSEnvSuite.TearDownTest(c)
+	s.IsolationSuite.TearDownTest(c)
 }
 
 func (s *ApplicationSuite) TestSetCharmStorageConstraints(c *gc.C) {
@@ -113,8 +137,9 @@ func (s *ApplicationSuite) TestSetCharmStorageConstraints(c *gc.C) {
 	})
 	c.Assert(err, jc.ErrorIsNil)
 	s.backend.CheckCallNames(c, "ModelTag", "Application", "Charm")
-	s.application.CheckCallNames(c, "SetCharm")
-	s.application.CheckCall(c, 0, "SetCharm", state.SetCharmConfig{
+	app := s.backend.applications["postgresql"].(*mockApplication)
+	app.CheckCallNames(c, "SetCharm")
+	app.CheckCall(c, 0, "SetCharm", state.SetCharmConfig{
 		Charm: &state.Charm{},
 		StorageConstraints: map[string]state.StorageConstraints{
 			"a": {},
@@ -133,9 +158,10 @@ func (s *ApplicationSuite) TestSetCharmConfigSettings(c *gc.C) {
 	})
 	c.Assert(err, jc.ErrorIsNil)
 	s.backend.CheckCallNames(c, "ModelTag", "Application", "Charm")
-	s.charm.CheckCallNames(c, "Config")
-	s.application.CheckCallNames(c, "SetCharm")
-	s.application.CheckCall(c, 0, "SetCharm", state.SetCharmConfig{
+	s.backend.charm.CheckCallNames(c, "Config")
+	app := s.backend.applications["postgresql"].(*mockApplication)
+	app.CheckCallNames(c, "SetCharm")
+	app.CheckCall(c, 0, "SetCharm", state.SetCharmConfig{
 		Charm:          &state.Charm{},
 		ConfigSettings: charm.Settings{"stringOption": "value"},
 	})
@@ -152,9 +178,10 @@ postgresql:
 	})
 	c.Assert(err, jc.ErrorIsNil)
 	s.backend.CheckCallNames(c, "ModelTag", "Application", "Charm")
-	s.charm.CheckCallNames(c, "Config")
-	s.application.CheckCallNames(c, "SetCharm")
-	s.application.CheckCall(c, 0, "SetCharm", state.SetCharmConfig{
+	s.backend.charm.CheckCallNames(c, "Config")
+	app := s.backend.applications["postgresql"].(*mockApplication)
+	app.CheckCallNames(c, "SetCharm")
+	app.CheckCall(c, 0, "SetCharm", state.SetCharmConfig{
 		Charm:          &state.Charm{},
 		ConfigSettings: charm.Settings{"stringOption": "value"},
 	})
@@ -182,9 +209,9 @@ func (s *ApplicationSuite) TestDestroyRelationRelationNotFound(c *gc.C) {
 }
 
 func (s *ApplicationSuite) TestBlockRemoveDestroyRelation(c *gc.C) {
-	s.blockChecker.SetErrors(errors.New("foo"))
+	s.blockChecker.SetErrors(errors.New("postgresql"))
 	err := s.api.DestroyRelation(params.DestroyRelation{Endpoints: []string{"a", "b"}})
-	c.Assert(err, gc.ErrorMatches, "foo")
+	c.Assert(err, gc.ErrorMatches, "postgresql")
 	s.blockChecker.CheckCallNames(c, "RemoveAllowed")
 	s.backend.CheckCallNames(c, "ModelTag")
 	s.relation.CheckNoCalls(c)
@@ -193,7 +220,7 @@ func (s *ApplicationSuite) TestBlockRemoveDestroyRelation(c *gc.C) {
 func (s *ApplicationSuite) TestDestroyApplication(c *gc.C) {
 	results, err := s.api.DestroyApplication(params.Entities{
 		Entities: []params.Entity{
-			{Tag: "application-foo"},
+			{Tag: "application-postgresql"},
 		},
 	})
 	c.Assert(err, jc.ErrorIsNil)
@@ -201,8 +228,8 @@ func (s *ApplicationSuite) TestDestroyApplication(c *gc.C) {
 	c.Assert(results.Results[0], jc.DeepEquals, params.DestroyApplicationResult{
 		Info: &params.DestroyApplicationInfo{
 			DestroyedUnits: []params.Entity{
-				{Tag: "unit-foo-0"},
-				{Tag: "unit-foo-1"},
+				{Tag: "unit-postgresql-0"},
+				{Tag: "unit-postgresql-1"},
 			},
 			DestroyedStorage: []params.Entity{
 				{Tag: "storage-pgdata-0"},
@@ -212,10 +239,10 @@ func (s *ApplicationSuite) TestDestroyApplication(c *gc.C) {
 }
 
 func (s *ApplicationSuite) TestDestroyApplicationNotFound(c *gc.C) {
-	s.backend.application = nil
+	delete(s.backend.applications, "postgresql")
 	results, err := s.api.DestroyApplication(params.Entities{
 		Entities: []params.Entity{
-			{Tag: "application-foo"},
+			{Tag: "application-postgresql"},
 		},
 	})
 	c.Assert(err, jc.ErrorIsNil)
@@ -223,7 +250,7 @@ func (s *ApplicationSuite) TestDestroyApplicationNotFound(c *gc.C) {
 	c.Assert(results.Results[0], jc.DeepEquals, params.DestroyApplicationResult{
 		Error: &params.Error{
 			Code:    params.CodeNotFound,
-			Message: `application "foo" not found`,
+			Message: `application "postgresql" not found`,
 		},
 	})
 }
@@ -231,8 +258,8 @@ func (s *ApplicationSuite) TestDestroyApplicationNotFound(c *gc.C) {
 func (s *ApplicationSuite) TestDestroyUnit(c *gc.C) {
 	results, err := s.api.DestroyUnit(params.Entities{
 		Entities: []params.Entity{
-			{Tag: "unit-foo-0"},
-			{Tag: "unit-foo-1"},
+			{Tag: "unit-postgresql-0"},
+			{Tag: "unit-postgresql-1"},
 		},
 	})
 	c.Assert(err, jc.ErrorIsNil)
@@ -248,218 +275,170 @@ func (s *ApplicationSuite) TestDestroyUnit(c *gc.C) {
 	}})
 }
 
-type mockBackend struct {
-	application.Backend
-	testing.Stub
-	application            *mockApplication
-	charm                  *mockCharm
-	endpoints              *[]state.Endpoint
-	relation               *mockRelation
-	unitStorageAttachments map[string][]state.StorageAttachment
-	storageInstances       map[string]*mockStorage
+func (s *ApplicationSuite) TestConsumeRequiresFeatureFlag(c *gc.C) {
+	s.SetFeatureFlags()
+	_, err := s.api.Consume(params.ConsumeApplicationArgs{})
+	c.Assert(err, gc.ErrorMatches, `set "cross-model" feature flag to enable consuming remote applications`)
 }
 
-func (b *mockBackend) ModelTag() names.ModelTag {
-	b.MethodCall(b, "ModelTag")
-	b.PopNoErr()
-	return coretesting.ModelTag
-}
-
-func (b *mockBackend) RemoteApplication(name string) (*state.RemoteApplication, error) {
-	b.MethodCall(b, "RemoteApplication", name)
-	return nil, errors.NotFoundf("remote application %q", name)
-}
-
-func (b *mockBackend) Application(name string) (application.Application, error) {
-	b.MethodCall(b, "Application", name)
-	if err := b.NextErr(); err != nil {
-		return nil, err
+func (s *ApplicationSuite) TestConsumeIdempotent(c *gc.C) {
+	for i := 0; i < 2; i++ {
+		results, err := s.api.Consume(params.ConsumeApplicationArgs{
+			Args: []params.ConsumeApplicationArg{{
+				ApplicationOffer: params.ApplicationOffer{
+					SourceModelTag:         coretesting.ModelTag.String(),
+					OfferName:              "hosted-mysql",
+					ApplicationDescription: "a database",
+					Endpoints:              []params.RemoteEndpoint{{Name: "database", Interface: "mysql", Role: "provider"}},
+					OfferURL:               "othermodel.hosted-mysql",
+				},
+			}},
+		})
+		c.Assert(err, jc.ErrorIsNil)
+		c.Assert(results.Results, gc.HasLen, 1)
+		c.Assert(results.Results[0].Error, gc.IsNil)
 	}
-	if b.application != nil {
-		return b.application, nil
+	obtained, ok := s.backend.remoteApplications["hosted-mysql"]
+	c.Assert(ok, jc.IsTrue)
+	c.Assert(obtained, jc.DeepEquals, &mockRemoteApplication{
+		name:           "hosted-mysql",
+		sourceModelTag: coretesting.ModelTag,
+		offerName:      "hosted-mysql",
+		offerURL:       "othermodel.hosted-mysql",
+		endpoints: []state.Endpoint{
+			{ApplicationName: "hosted-mysql", Relation: charm.Relation{Name: "database", Interface: "mysql", Role: "provider"}}},
+	})
+}
+
+func (s *ApplicationSuite) TestConsumeIncludesSpaceInfo(c *gc.C) {
+	s.env.spaceInfo = &environs.ProviderSpaceInfo{
+		CloudType: "grandaddy",
+		ProviderAttributes: map[string]interface{}{
+			"thunderjaws": 1,
+		},
+		SpaceInfo: network.SpaceInfo{
+			Name:       "yourspace",
+			ProviderId: "juju-space-myspace",
+			Subnets: []network.SubnetInfo{{
+				CIDR:              "5.6.7.0/24",
+				ProviderId:        "juju-subnet-1",
+				AvailabilityZones: []string{"az1"},
+			}},
+		},
 	}
-	return nil, errors.NotFoundf("application %q", name)
-}
 
-func (b *mockBackend) Unit(name string) (application.Unit, error) {
-	b.MethodCall(b, "Unit", name)
-	if err := b.NextErr(); err != nil {
-		return nil, err
+	results, err := s.api.Consume(params.ConsumeApplicationArgs{
+		Args: []params.ConsumeApplicationArg{{
+			ApplicationAlias: "beirut",
+			ApplicationOffer: params.ApplicationOffer{
+				SourceModelTag:         coretesting.ModelTag.String(),
+				OfferName:              "hosted-mysql",
+				ApplicationDescription: "a database",
+				Endpoints:              []params.RemoteEndpoint{{Name: "server", Interface: "mysql", Role: "provider"}},
+				OfferURL:               "othermodel.hosted-mysql",
+				Bindings:               map[string]string{"server": "myspace"},
+				Spaces: []params.RemoteSpace{
+					{
+						CloudType:  "grandaddy",
+						Name:       "myspace",
+						ProviderId: "juju-space-myspace",
+						ProviderAttributes: map[string]interface{}{
+							"thunderjaws": 1,
+						},
+						Subnets: []params.Subnet{{
+							CIDR:       "5.6.7.0/24",
+							ProviderId: "juju-subnet-1",
+							Zones:      []string{"az1"},
+						}},
+					},
+				},
+			},
+		}},
+	})
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(results.Results, gc.HasLen, 1)
+	c.Assert(results.Results[0].Error, gc.IsNil)
+	c.Assert(results.Results[0].LocalName, gc.Equals, "beirut")
+
+	obtained, ok := s.backend.remoteApplications["beirut"]
+	c.Assert(ok, jc.IsTrue)
+	endpoints, err := obtained.Endpoints()
+	c.Assert(err, jc.ErrorIsNil)
+	epNames := make([]string, len(endpoints))
+	for i, ep := range endpoints {
+		epNames[i] = ep.Name
 	}
-	if b.application != nil {
-		for _, u := range b.application.units {
-			if u.tag.Id() == name {
-				return &u, nil
-			}
-		}
-	}
-	return nil, errors.NotFoundf("unit %q", name)
+	c.Assert(epNames, jc.SameContents, []string{"server"})
+	c.Assert(obtained.Bindings(), jc.DeepEquals, map[string]string{"server": "myspace"})
+	c.Assert(obtained.Spaces(), jc.DeepEquals, []state.RemoteSpace{{
+		CloudType:  "grandaddy",
+		Name:       "myspace",
+		ProviderId: "juju-space-myspace",
+		ProviderAttributes: map[string]interface{}{
+			"thunderjaws": 1,
+		},
+		Subnets: []state.RemoteSubnet{{
+			CIDR:              "5.6.7.0/24",
+			ProviderId:        "juju-subnet-1",
+			AvailabilityZones: []string{"az1"},
+		}},
+	}})
 }
 
-func (b *mockBackend) Charm(curl *charm.URL) (application.Charm, error) {
-	b.MethodCall(b, "Charm", curl)
-	if err := b.NextErr(); err != nil {
-		return nil, err
-	}
-	if b.charm != nil {
-		return b.charm, nil
-	}
-	return nil, errors.NotFoundf("charm %q", curl)
-}
+// TODO(wallyworld) - re-implement when OfferDetails is done
+//func (s *ApplicationSuite) TestConsumeRejectsEndpoints(c *gc.C) {
+//	results, err := s.api.Consume(params.ConsumeApplicationArgs{
+//		Args: []params.ConsumeApplicationArg{{ApplicationURL: "othermodel.application:db"}},
+//	})
+//	c.Assert(err, jc.ErrorIsNil)
+//	c.Assert(results.Results, gc.HasLen, 1)
+//	c.Assert(results.Results[0].Error != nil, jc.IsTrue)
+//	c.Assert(results.Results[0].Error.Message, gc.Equals, `remote application "othermodel.application:db" shouldn't include endpoint`)
+//}
+//
+//func (s *ApplicationSuite) TestConsumeNoPermission(c *gc.C) {
+//	s.setupOffer()
+//	s.mockBackend.users.Add("someone")
+//	user := names.NewUserTag("someone")
+//	offer := names.NewApplicationOfferTag("hosted-mysql")
+//	err := s.mockBackend.CreateOfferAccess(offer, user, permission.NoAccess)
+//	c.Assert(err, jc.ErrorIsNil)
+//
+//	targetModelTag := s.setupTargetModel()
+//
+//	s.authorizer.Tag = names.NewUserTag("someone")
+//	results, err := s.api.Consume(params.ConsumeApplicationArgs{
+//		Args: []params.ConsumeApplicationArg{{
+//			SourceModelTag:         testing.ModelTag.String(),
+//			OfferName:              "hosted-mysql",
+//			ApplicationDescription: "a database",
+//			Endpoints:              []params.RemoteEndpoint{{Name: "database", Interface: "mysql", Role: "provider"}},
+//			OfferURL:               "othermodel.hosted-mysql",
+//			ApplicationAlias:       "mysql",
+//			TargetModelTag:         targetModelTag.String(),
+//		}},
+//	})
+//	c.Assert(err, jc.ErrorIsNil)
+//	c.Assert(results.Results, gc.HasLen, 1)
+//	c.Assert(results.Results[0].Error, gc.ErrorMatches, ".*permission denied.*")
+//}
+//
+//func (s *ApplicationSuite) TestConsumeWithPermission(c *gc.C) {
+//	st := s.mockStatePool.st[testing.ModelTag.Id()]
+//	st.(*mockBackend).users.Add("postgresqlbar")
 
-func (b *mockBackend) InferEndpoints(endpoints ...string) ([]state.Endpoint, error) {
-	b.MethodCall(b, "InferEndpoints", endpoints)
-	if err := b.NextErr(); err != nil {
-		return nil, err
-	}
-	if b.endpoints != nil {
-		return *b.endpoints, nil
-	}
-	return nil, errors.Errorf("no relations found")
-}
-
-func (b *mockBackend) EndpointsRelation(endpoints ...state.Endpoint) (application.Relation, error) {
-	b.MethodCall(b, "EndpointsRelation", endpoints)
-	if err := b.NextErr(); err != nil {
-		return nil, err
-	}
-	if b.relation != nil {
-		return b.relation, nil
-	}
-	return nil, errors.NotFoundf("relation")
-}
-
-func (b *mockBackend) UnitStorageAttachments(tag names.UnitTag) ([]state.StorageAttachment, error) {
-	b.MethodCall(b, "UnitStorageAttachments", tag)
-	if err := b.NextErr(); err != nil {
-		return nil, err
-	}
-	return b.unitStorageAttachments[tag.Id()], nil
-}
-
-func (b *mockBackend) StorageInstance(tag names.StorageTag) (state.StorageInstance, error) {
-	b.MethodCall(b, "StorageInstance", tag)
-	if err := b.NextErr(); err != nil {
-		return nil, err
-	}
-	s, ok := b.storageInstances[tag.Id()]
-	if !ok {
-		return nil, errors.NotFoundf("storage %s", tag.Id())
-	}
-	return s, nil
-}
-
-type mockApplication struct {
-	application.Application
-	testing.Stub
-	units []mockUnit
-}
-
-func (a *mockApplication) AllUnits() ([]application.Unit, error) {
-	a.MethodCall(a, "AllUnits")
-	if err := a.NextErr(); err != nil {
-		return nil, err
-	}
-	units := make([]application.Unit, len(a.units))
-	for i := range a.units {
-		units[i] = &a.units[i]
-	}
-	return units, nil
-}
-
-func (a *mockApplication) SetCharm(cfg state.SetCharmConfig) error {
-	a.MethodCall(a, "SetCharm", cfg)
-	return a.NextErr()
-}
-
-func (a *mockApplication) Destroy() error {
-	a.MethodCall(a, "Destroy")
-	return a.NextErr()
-}
-
-type mockCharm struct {
-	application.Charm
-	testing.Stub
-	config *charm.Config
-}
-
-func (c *mockCharm) Config() *charm.Config {
-	c.MethodCall(c, "Config")
-	c.PopNoErr()
-	return c.config
-}
-
-type mockBlockChecker struct {
-	testing.Stub
-}
-
-func (c *mockBlockChecker) ChangeAllowed() error {
-	c.MethodCall(c, "ChangeAllowed")
-	return c.NextErr()
-}
-
-func (c *mockBlockChecker) RemoveAllowed() error {
-	c.MethodCall(c, "RemoveAllowed")
-	return c.NextErr()
-}
-
-type mockRelation struct {
-	application.Relation
-	testing.Stub
-}
-
-func (r *mockRelation) Destroy() error {
-	r.MethodCall(r, "Destroy")
-	return r.NextErr()
-}
-
-type mockUnit struct {
-	application.Unit
-	testing.Stub
-	tag names.UnitTag
-}
-
-func (u *mockUnit) UnitTag() names.UnitTag {
-	return u.tag
-}
-
-func (u *mockUnit) IsPrincipal() bool {
-	u.MethodCall(u, "IsPrincipal")
-	u.PopNoErr()
-	return true
-}
-
-func (u *mockUnit) Destroy() error {
-	u.MethodCall(u, "Destroy")
-	return u.NextErr()
-}
-
-type mockStorageAttachment struct {
-	state.StorageAttachment
-	testing.Stub
-	unit    names.UnitTag
-	storage names.StorageTag
-}
-
-func (a *mockStorageAttachment) Unit() names.UnitTag {
-	return a.unit
-}
-
-func (a *mockStorageAttachment) StorageInstance() names.StorageTag {
-	return a.storage
-}
-
-type mockStorage struct {
-	state.StorageInstance
-	testing.Stub
-	tag   names.StorageTag
-	owner names.Tag
-}
-
-func (a *mockStorage) StorageTag() names.StorageTag {
-	return a.tag
-}
-
-func (a *mockStorage) Owner() (names.Tag, bool) {
-	return a.owner, a.owner != nil
-}
+//_, err := s.otherModel.AddUser("someone", "spmeone", "secret", "admin")
+//c.Assert(err, jc.ErrorIsNil)
+//apiUser := names.NewUserTag("someone")
+//err = s.otherModel.CreateOfferAccess(
+//	names.NewApplicationOfferTag("hosted-mysql"), apiUser, permission.ConsumeAccess)
+//s.authorizer.Tag = apiUser
+//results, err := s.api.Consume(params.ConsumeApplicationArgs{
+//	Args: []params.ConsumeApplicationArg{
+//		{ApplicationURL: "admin/othermodel.hosted-mysql"},
+//	},
+//})
+//c.Assert(err, jc.ErrorIsNil)
+//c.Assert(results.Results, gc.HasLen, 1)
+//c.Assert(results.Results[0].Error, gc.IsNil)
+//}
